arm64: configs: enable CONFIG_DYNAMIC_FTRACE
Enabling the "CONFIG_DYNAMIC_FTRACE" macro which was disabled earlier
due to GVS failures. Also, disabling the "CONFIG_ARM64_PTR_AUTH" feature
in this change as the failures are coming due to the pointer Authentication
feature in both L4T and Auto platforms but for different reasons.
In L4T, "PTR_AUTH" is resulting in timeouts due to perf drop in some tests.
In Auto, this macro seems to be catching some issue. After Auto issue is
fixed, we can enable "ARM64_PTR_AUTH" feature in debug build if we need
and increase the timeout period of tests impacted by perf drop.

Bug 3341920

Signed-off-by: Sumit Gupta <sumitg@nvidia.com>
Change-Id: If8b705b90cdae0b4010dba5d5bb3a1dffd81e239
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2565628
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>