PCI: dwc: Fix incorrect ATU window assignment
commit 40f14cecc8c3 ("PCI: dwc: Support multiple ATU memory regions")
makes the assumption that if a platform implements its own way of
translating configuration space accesses, then, it must have also
implemented its own way of translating memory/IO accesses which need
not be true. It is perfectly fine for a platform to implement ECAM method
to translate configuration space accesses but rely on DWC's ATU windows
for memory/IO translations. This change addresses that issue by skipping
ATU usage only for configuration translations for a given platform which
has its own implementation for the same.

Fixes: 40f14cecc8c3 ("PCI: dwc: Support multiple ATU memory regions")
Change-Id: I6542e36287aafe81f64447b515735705eae83eaf
Signed-off-by: Vidya Sagar <vidyas@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2572773
Tested-by: Manikanta Maddireddy <mmaddireddy@nvidia.com>
Tested-by: Ashish Mhetre <amhetre@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Manikanta Maddireddy <mmaddireddy@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit