dts: t23x: Adding AZA clock ids.
This change adds the AZA_BIT and AZA_2XBIT clocks as part of
display's DT clock entries. RM needs to enable these clocks
during state init in order to read/write HDA registers when
initializing the HDACODEC object.

Exposing clock string namespace attached to clock ids in
BPMP to the RM to use it for the CLOCK and PLL programming
for Read/Write.

Jira Id TDS-6554

Change-Id: Ibcf65c0050ccb809fd343d2449ea0deb33870ff1
Signed-off-by: Gaurav Madheshia <gmadheshia@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/soc/t23x/+/2573018
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>