video: tegra: nvmap: Correct free memory value
NvRmMemQueryHeapParams API return freeram as free memory for IOVMM
memory. Free memory only consider the free pages and does not consider
any cached memory which can be reclaimed during memory pressure.
Available memory is the kernel estimate for the allocatable memory, so
it is a correct value to be reported by this API. But the CMA memory is
not allocatable by NvMap for regular allocations and it is part of the
reported available memory. Hence subtract it from available memory value
NvMap overallocate pages by 1/16th in case of page coloring. So subtract
this extra value as well.

Bug 200724226

Change-Id: I464a7ef41a4a0bed2a53c2ec0f77771382e88bb7
Signed-off-by: Ketan Patil <ketanp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2583714
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>