t234: uncore perfmon driver fix cpu assignment
On system wide counting, the perf kernel driver will try
to count an uncore event on the cpus that are defined on the
pmu cpumask attribute. The uncore driver doesnt define cpumask
attribute, so the kernel will try to count uncore event on all
cpus.
Since uncore events are system-wide and not mapped to a particular
core, the uncore driver always chooses core 0 to count the uncore events.
It does this by checking event->core attribute and ignore the
ones that are non-zero.
However, the event_init procedure overwrites this attribute to 0.
This makes the event_init call from other cpus is "converted" to cpu 0.
The fix is to remove this assignment to event-core attribute during
event_init.

JIRA MSST-380

Change-Id: Ibe8e81f2b81dc19ccc5e464a78e2570f9eeb129e
Signed-off-by: Besar Wicaksono <bwicaksono@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2584807
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Richard Wiley <rwiley@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>