phy: tegra: xusb: fix use-after-free issue
For the OTG port, it will assign the phy dev to usb-phy dev and
use the port dev driver as the dev driver of usb-phy.

When we try to destroy the port dev, it will destroy its dev driver
as well. But we did not remove the reference from usb-phy dev. This
might cause the use-after-free issue in KASAN.

Bug 200750395

Change-Id: I11707b4aae2d7f7f9ac9f3cf9c83117496914b9d
Signed-off-by: EJ Hsu <ejh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2590676
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>