video: tegra: nvmap: Fix available memory check
When allocation is requested from the heap which is not present then
is_nvmap_memory_available() call is returning true after iterating through
all the available heaps and allocation is proceeding further.
Fix this by returning false from that function is heap is not present.

Bug 200769675

Change-Id: I18aed3bfae04c8bcf1fdbf0814ceaceb09228797
Signed-off-by: Ashish Mhetre <amhetre@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2597236
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Puneet Saxena <puneets@nvidia.com>
Reviewed-by: Ketan Patil <ketanp@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>