arch: arm64: configs: TOT config uploaded
Configuring the kernel first time reorders the configs even if
there are no changes. This becomes a problem when going through
git diff. So to avoid that problem kernel config is done on TOT
just to reorder the configs. So eventually when other config changes
are made git diff can be easily interpreted.

Bug 200755555

Change-Id: Ifb68491e144ddbf26320b46d9212dcc6e07f3512
Signed-off-by: Sahil Mukund Patki <spatki@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2597256
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>