tegra: soc_hwpm: Fix debugfs compilation errors
The member debugfs_root is only defined in struct tegra_soc_hwpm when
debugfs is enabled. So its use when debugfs is disabled leads to
compilation errors which is seen in function tegra_soc_hwpm_debugfs_init.
This function when debugfs disabled just sets the debugfs_root to NULL.
This has been removed to fix the compilation error.

Bug 200755555

Change-Id: I6e4fa7e330dbe468ba216e98d29d58877e250fa4
Signed-off-by: Sahil Mukund Patki <spatki@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2616087
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Deepak Nibade <dnibade@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit