ASoC: rt5659: Fix duplicate prefix application
Following error prints are seen with kernel-5.10.65:
  rt5659 7-001a: ASoC: unknown pin CVB-RT LDO2
  rt5659 7-001a: ASoC: unknown pin CVB-RT MICBIAS1
  rt5659 7-001a: ASoC: unknown pin CVB-RT Mic Det Power
  rt5659 7-001a: ASoC: DAPM unknown pin CVB-RT LDO2
  rt5659 7-001a: ASoC: DAPM unknown pin CVB-RT MICBIAS1
  rt5659 7-001a: ASoC: DAPM unknown pin CVB-RT Mic Det Power

This is seen on Xaveir SLT platforms. This is expected to be seen
with Galen and Concord as well with super-module usage.

The commit ae4fc532244b ("ASoC: dapm: use component prefix when
checking widget names") present in kernel-5.10.65 is causing this
failure. The primary reason this happens is, DAPM core now uses
component prefix for widget checks and the prefix application
already present in codec driver results in duplication of prefix.
So this means prefix is getting applied twice, once from codec
driver and once from DAPM core.

Fix this by removing prefix handling in codec driver and let this
be managed by DAPM core itself.

Bug 200712834

Change-Id: I0fc655f0646a14e9391df24ec34d87cc87605f38
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2619407
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>