thermal: tegra-bpmp: fix offline sensor retval
Returning an error code from the get_temp on boot causes the thermal
framwork to skip programming of trip points. Even after the sensor
comes back online since no trips were ever programmed the thermal
framework is not notified about the unrailgating event which causes
the trips to never be programmed.

Fix this issue by returning a large negative value that is not
INT32_MIN (which is treated specially by the thermal framework) so
that trips are registered and the framework can get notified when
the sensor comes back online.

Bug 3402180
Bug 200780375
Bug 3400836

Change-Id: I8aa2889d4ff029cf5d0c7983ff1f6ba9f0f7d0b5
Signed-off-by: Srikar Srimath Tirumala <srikars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2624033
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Timo Alho <talho@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit