soc/tegra: bpmp: Mark XUSB domains as ACTIVE_WAKEUP on T234
On Tegra234, XUSB domains need to be left in powered state when
entering system suspend to ensure that upon resume wakeup-related
information is available. To that end, mark the domains with
GENPD_FLAG_ACTIVE_WAKEUP. This in combination with the XUSB device
being marked as a wakeup source ensures the power domains are
left powered when entering system suspend.

Change-Id: I7ea269583dcfa296c402c81b250b004f930cf9ba
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2631189
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Stefan Kristiansson <stefank@nvidia.com>
Reviewed-by: Timo Alho <talho@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: Henry Lin <henryl@nvidia.com>
Tested-by: Timo Alho <talho@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit