gpu: nvgpu: evaluate srctree.nvgpu immediately
The definition of srctree.nvgpu for external modules is specified as a
function of $(MAKEFILE_LIST). This variable gets extended also when
files are included, so any include directives in the makefile affect the
path because srctree.nvgpu is assigned with "?=" which is lazy like "=".

To avoid problems with the lazy init, use ":=" surrounded with an
explicit check to initialize srctree.nvgpu only when undefined.

Jira NVGPU-6427

Change-Id: Ic9847528d0da73f06bc7e421130c9c54247caadb
Signed-off-by: Konsta Hölttä <kholtta@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvgpu/+/2632925
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>