video: tegra: nvmap: Add null pointer check for device node
If VPR resize is not enabled then VPR dt node is removed by BL code
because of which memory-region node in tegra-carveouts has vpr phandle
but it becomes an empty entry with no arguments. But
of_phandle_iterator_next does not return error in such cases, because
of which, null device node would be passed to of_reserved_mem_lookup
and results in null pointer dereference. Fix this by adding null pointer
check for the device node.

Bug 3415117

Change-Id: I8feaf1d3348f9e2617248dde4b6925c9102475a5
Signed-off-by: Ketan Patil <ketanp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2634148
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Ashish Mhetre <amhetre@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit